Huge commit, my bad.
Changing the whole Aiming component structure, making it more BP based and ran into a lot of crashes.
Now using assertions, ensure() on this case, allowed me to see what is going on with the null pointers.

Using BlueprintImplementableEvent to set the aiming component reference in BP

Still in process.